feat: add publishing app functions [WIP]
NOTE: The artifact upload needs to be authentication aware
